otel-lab/requirements.txt
flask
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
